+++ apps/official/src/constants/index.ts
@@ -1,1 +1,8 @@
 export * from "./packages";
+
+const isDev = process.env.NODE_ENV !== "production";
+
+export default {
+	isDev,
+	localStorageKeyPrefix: "hike-docs",
+};