+++ packages/hooks/src/index.ts
@@ -1,5 +1,2 @@
 export * from "./use-hike";
 export * from "./use-model";
-export * from "./use-query";
-export * from "./use-mount";
-export * from "./use-unmount";