+++ packages/hooks/src/use-model/index.ts
@@ -1,2 +1,10 @@
-export * from "./vanilla.ts";
-export * from "./react.ts";
+import * as vanillaUtils from "./vanilla/utils";
+import * as reactUtils from "./react/utils";
+
+export * from "./vanilla";
+export * from "./react";
+
+export const modelUtils = {
+	...vanillaUtils,
+	...reactUtils,
+};