+++ packages/hooks/src/use-model/react/utils/useProxy.ts
@@ -1,5 +1,5 @@
 import { useLayoutEffect } from "react";
-import { useModel } from "../../react.ts";
+import { useModel } from "../../react";
 
 const DUMMY_SYMBOL = Symbol();