+++ packages/hooks/src/use-model/utils.ts
@@ -1,2 +1,0 @@
-export * from "./vanilla/utils.ts";
-export * from "./react/utils.ts";