+++ packages/hooks/src/use-model/vanilla/utils.ts
@@ -1,6 +1,6 @@
-export { subscribeKey } from "./utils/subscribeKey.ts";
-export { watch } from "./utils/watch.ts";
-export { devtools } from "./utils/devtools.ts";
-export { deepClone } from "./utils/deepClone.ts";
-export { proxySet } from "./utils/proxySet.ts";
-export { proxyMap } from "./utils/proxyMap.ts";
+export { subscribeKey } from "./utils/subscribeKey";
+export { watch } from "./utils/watch";
+export { devtools } from "./utils/devtools";
+export { deepClone } from "./utils/deepClone";
+export { proxySet } from "./utils/proxySet";
+export { proxyMap } from "./utils/proxyMap";