+++ packages/hooks/src/use-model/vanilla/utils/proxyMap.ts
@@ -1,4 +1,4 @@
-import { proxy } from "../../vanilla.ts";
+import { proxy } from "../../vanilla";
 
 type KeyValRecord<K, V> = [key: K, value: V];