+++ packages/jvm/src/index.ts
@@ -1,0 +1,3 @@
+export * from "./jvm-module-options.interface";
+export * from "./jvm.module";
+export * from "./jvm.service";