+++ packages/jvm/src/jvm-module-options.interface.ts
@@ -1,0 +1,4 @@
+export interface JvmModuleOptions {
+	baseDir?: string;
+	dependencies?: string[];
+}