+++ packages/prisma/src/prisma-module-options.interface.ts
@@ -1,7 +1,1 @@
-import { Prisma } from "@prisma/client";
-
-export interface PrismaModuleOptions
-	extends Prisma.Subset<
-		Prisma.PrismaClientOptions,
-		Prisma.PrismaClientOptions
-	> {}
+export type PrismaModuleOptions = {};