+++ packages/aliyun/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @hikestack/aliyun
 
+## 1.1.1
+
+### Patch Changes
+
+- fix: injection AliyunOssService
+
 ## 1.1.0
 
 ### Minor Changes