+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @hikestack/hooks
 
+## 2.2.2
+
+### Patch Changes
+
+- fix: path alias error
+
+## 2.2.1
+
+### Patch Changes
+
+- fix: export files
+
+## 2.2.0
+
+### Minor Changes
+
+- feat: update export files
+
 ## 2.1.0
 
 ### Minor Changes
@@ -16,9 +34,6 @@
 
 ### Minor Changes
 
-- fix: npm ignore
-- first release
-- 1aeea38: first release
-- add new some hooks
+- add some hooks
 - fix: npm ignore
 - first release